refactor: consolidate GLTF input resolution into a dedicated function
